feat: bigDecimal 추가 및 ktlint 수정

+++ src/main/kotlin/com/my/sparta/concert/aggregate/user/adapter/inbound/web/interfaces/request/UserChargeMoneyRequest.kt
@@ -1,10 +1,11 @@
 package com.my.sparta.concert.aggregate.user.adapter.inbound.web.interfaces.request
 
 import lombok.Data
+import java.math.BigDecimal
 
 @Data
 class UserChargeMoneyRequest(
     val userId: String,
-    val money: Double,
+    val money: BigDecimal,
     val paymentType: String,
 )
